Estimate Down Levels

Non-geographic partner function to geo_estimate_down. Allows users to estimate down without the costly matching operation if they've already matched.

estimate_down(wts, value, group)

Arguments

wts
numeric vector. Defaults to 1. Typically population or VAP, as a weight to give each precinct.

value
numeric vector. Defaults to 1. Typically electoral outcomes, as a value to estimate down into blocks.

group
matches of length(wts) that correspond to row indices of value. Often, this input is the output of geo_match.

Value

numeric vector with each value split by weight

Examples

if (FALSE) {
data("va18sub")
va18sub <- va18sub %>% filter(COUNTYFP == '087')
block <- create_block_table(state = 'VA', county = '087')  
groups <- geo_match(block, va18sub)
disagg <- estimate_down(wts = block$VAP, value = va18sub$G18USSRSTE, group = groups)
}