Estimate Plans from a Block Assignment File to Voting Districts

District lines are often provided at the census block level, but analyses often occur at the voting district level. This provides a simple way to estimate the block level to the voting district level.

baf_to_vtd(baf, plan_name, GEOID = "GEOID", year = 2020)

Arguments

baf

a tibble representing a block assignment file.

plan_name

character. Name of column in baf which corresponds to the districts.

GEOID

character. Name of column which corresponds to each block's GEOID, sometimes called "BLOCKID". Default is 'GEOID'.

year

the decade to request, either 2010 or 2020. Default is 2020.

Value

a tibble with a vtd-level assignment file

Details

If a voting district is split between blocks, this currently uses the most common district.

Examples

# Not guaranteed to reach download from redistrict2020.org
if (FALSE) { # \dontrun{
# download and read baf ----
url <- paste0('https://github.com/PlanScore/Redistrict2020/', 
              'raw/main/files/DE-2021-01/DE_SLDU_bef.zip')
tf <- tempfile('.zip')
utils::download.file(url, tf)
utils::unzip(tf, exdir = dirname(tf))
baf <- readr::read_csv(
  file = paste0(dirname(tf), '/DE_SLDU_bef.csv'),
  col_types = 'ci'
)
names(baf) <- c('GEOID', 'ssd_20')

# convert to vtd level ----
baf_to_vtd(baf = baf, plan_name = 'ssd_20', 'GEOID')
} # }