Get historical United States Congressional District Shapefiles

Data sourced from the United States Congressional District Shapefiles, primarily hosted at https://cdmaps.polisci.ucla.edu/. Files are fetched through the GitHub repository at https://github.com/JeffreyBLewis/congressional-district-boundaries.

get_lewis(state, congress)

Arguments

state

two letter state abbreviation

congress

congress number, from 1 to 114.

Value

a sf tibble of the congressional district boundaries

References

Jeffrey B. Lewis, Brandon DeVine, Lincoln Pitcher, and Kenneth C. Martis. (2013) Digital Boundary Definitions of United States Congressional Districts, 1789-2012. [Data file and code book]. Retrieved from https://cdmaps.polisci.ucla.edu on [date of download].

Examples

get_lewis(state = 'NM', congress = 111)
#> Data sourced from the United States Congressional District Shapefiles
#> <https://cdmaps.polisci.ucla.edu/>.
#> This message is displayed once per session.
#> Simple feature collection with 3 features and 6 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -109.0502 ymin: 31.3323 xmax: -103.002 ymax: 37.00023
#> Geodetic CRS:  WGS 84
#> # A tibble: 3 × 7
#>   startcong district statename  member           endcong id          
#> * <chr>     <chr>    <chr>      <list>           <chr>   <chr>       
#> 1 108       2        New Mexico <named list [5]> 112     035108112002
#> 2 108       3        New Mexico <named list [5]> 112     035108112003
#> 3 108       1        New Mexico <named list [5]> 112     035108112001
#> # ℹ 1 more variable: geometry <MULTIPOLYGON [°]>